[MISC] Move seqan3::views::to_char to seqan3/alphabet/views/
