+++ include/seqan3/range/views/all.hpp
@@ -13,6 +13,7 @@
 #pragma once
 
 #include <seqan3/alphabet/views/char_to.hpp>
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/enforce_random_access.hpp>
 #include <seqan3/range/views/get.hpp>
@@ -27,7 +28,6 @@
 #include <seqan3/range/views/take_line.hpp>
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/to_rank.hpp>
 #include <seqan3/range/views/translate.hpp>
 #include <seqan3/range/views/trim_quality.hpp>
