+++ test/snippet/io/structure_file/structure_file_input_record_iter.cpp
@@ -1,8 +1,8 @@
 #include <sstream>
 
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/io/structure_file/input.hpp>
-#include <seqan3/range/views/to_char.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
 
 auto input = R"(> S.cerevisiae_tRNA-PHE M10740/1-73
