+++ test/snippet/range/views/get.cpp
@@ -3,9 +3,9 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/quality/aliases.hpp> // includes seqan3::dna4q
 #include <seqan3/alphabet/quality/phred42.hpp>
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/range/views/get.hpp>
-#include <seqan3/range/views/to_char.hpp>
 
 int main()
 {
