+++ test/snippet/range/views/trim_quality_dna5q.cpp
@@ -4,8 +4,8 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/quality/aliases.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/range/views/trim_quality.hpp>
-#include <seqan3/range/views/to_char.hpp>
 
 int main()
 {
