+++ test/unit/alignment/matrix/edit_distance_matrix/trace_matrix_path_iterator_test.cpp
@@ -11,7 +11,7 @@
 
 #include <seqan3/alignment/matrix/detail/two_dimensional_matrix_iterator_concept.hpp>
 #include <seqan3/alignment/matrix/edit_distance_trace_matrix_full.hpp>
-#include <seqan3/range/views/to_char.hpp>
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/pretty_printing.hpp>
