+++ test/unit/range/views/trim_quality_test.cpp
@@ -10,9 +10,9 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/alphabet/quality/aliases.hpp>
+#include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/range/concept.hpp>
-#include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/trim_quality.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
